+++ animation_nodes/nodes/spline/spline_evaluation_base.py
@@ -14,6 +14,6 @@ class SplineEvaluationBase:
     parameterType: EnumProperty(name = "Parameter Type", default = "UNIFORM",
         items = parameterTypeItems, update = parameterTypeChanged)
 
-    resolution: IntProperty(name = "Resolution", min = 2, default = 5,
+    resolution: IntProperty(name = "Resolution", min = 2, default = 20,
         description = "Increase to have a more accurate evaluation if the type is set to Uniform",
         update = propertyChanged)